+++ src/main/java/com/andersmmg/lockandblock/config/ModConfigModel.java
@@ -19,13 +19,13 @@ public class ModConfigModel {
     @RangeConstraint(min = 1.0f, max = 10.0f)
     public float playerSensorRange = 4.0f;
     @RangeConstraint(min = 3, max = 30)
-    public int maxTripMineDistance = 7;
-    @RangeConstraint(min = 3, max = 30)
-    public int maxLaserSensorDistance = 10;
+    public int maxLaserDistance = 10;
     @RangeConstraint(min = 1.0f, max = 10.0f)
     public float redstoneLaserDamage = 3.0f;
-    @RangeConstraint(min = 1, max = 60)
+    @RangeConstraint(min = 1, max = 255)
     public int redstonePulseLength = 30;
+    public boolean laserParticles = false;
+    public boolean laserCustomRender = true;
 
     @SuppressWarnings("unused")
     public boolean allowTripMinesAir = true;